Refactor MenuItem into separate file
